Move icon-button to separate folder. Because sg doesn't show Icon sample.
