+++ front/src/app/components/common/icon-button/index.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+import { Icon, IIconProps } from 'app/components/common/icon';
+
+export const IconButton = (props: IIconProps) => (
+    <Icon {...props} tag="button" />
+);
